feat: creation of api data adapter
