+++ src/adapters/api-data/api-data-adapter.ts
@@ -1,0 +1,17 @@
+import { CenoteModel } from '../../models/CenotesTypes';
+import ReferenceModel from '../../models/ReferencesTypes';
+import { VariableModel } from '../../models/VariablesTypes';
+
+const classMap = {
+  cenotes: (data: CenoteModel) => new CenoteModel(data),
+  variables: (data: VariableModel) => new VariableModel(data),
+  references: (data: ReferenceModel) => new ReferenceModel(data),
+};
+
+export const dataAdapter = (classType: string, data: any) => {
+  const clazz = classMap[classType as keyof typeof classMap];
+  return data.content.map(
+    (cenote: CenoteModel & VariableModel & ReferenceModel) =>
+      clazz(cenote)
+  );
+};
